+++ fltk-sys/cfltk/src/cfl_text.cpp
@@ -1,13 +1,13 @@
 #include "cfl_text.h"
 #include <FL/Fl.H>
 
+#include "cfl_new.hpp"
 #include <FL/Fl_Image.H>
 #include <FL/Fl_Simple_Terminal.H>
 #include <FL/Fl_Text_Buffer.H>
 #include <FL/Fl_Text_Display.H>
 #include <FL/Fl_Text_Editor.H>
 #include <FL/Fl_Widget.H>
- #include "cfl_new.hpp"
 
 #define DISPLAY_DEFINE(widget)                                                                     \
     int widget##_text_font(const widget *self) {                                                   \
@@ -70,16 +70,15 @@
     void widget##_show_cursor(widget *self, int boolean) {                                         \
         LOCK(if (boolean) self->show_cursor(); else self->hide_cursor();)                          \
     }                                                                                              \
-    void *widget##_set_style_table_entry(widget *self, void *sbuff, unsigned int *color,           \
-                                         int *font, int *fontsz, int sz) {                         \
+    void widget##_set_style_table_entry(widget *self, void *sbuff, unsigned int *color, int *font, \
+                                        int *fontsz, int sz) {                                     \
         Fl_Text_Display::Style_Table_Entry *stable = new Fl_Text_Display::Style_Table_Entry[sz];   \
         if (!stable)                                                                               \
-            return NULL;                                                                           \
+            return;                                                                                \
         for (int i = 0; i < sz; ++i) {                                                             \
             stable[i] = {color[i], font[i], fontsz[i]};                                            \
         }                                                                                          \
         LOCK(self->highlight_data((Fl_Text_Buffer *)sbuff, stable, sz, 'A', 0, 0);)                \
-        return (void *)stable;                                                                     \
     }                                                                                              \
     void widget##_set_cursor_style(widget *self, int style) {                                      \
         LOCK(self->cursor_style(style);)                                                           \