+++ fltk-sys/src/text.rs
@@ -734,7 +734,7 @@ extern "C" {
         font: *mut libc::c_int,
         fontsz: *mut libc::c_int,
         sz: libc::c_int,
-    ) -> *mut libc::c_void;
+    );
 }
 extern "C" {
     pub fn Fl_Text_Display_set_cursor_style(arg1: *mut Fl_Text_Display, style: libc::c_int);
@@ -1173,7 +1173,7 @@ extern "C" {
         font: *mut libc::c_int,
         fontsz: *mut libc::c_int,
         sz: libc::c_int,
-    ) -> *mut libc::c_void;
+    );
 }
 extern "C" {
     pub fn Fl_Text_Editor_set_cursor_style(arg1: *mut Fl_Text_Editor, style: libc::c_int);
@@ -1745,7 +1745,7 @@ extern "C" {
         font: *mut libc::c_int,
         fontsz: *mut libc::c_int,
         sz: libc::c_int,
-    ) -> *mut libc::c_void;
+    );
 }
 extern "C" {
     pub fn Fl_Simple_Terminal_set_cursor_style(arg1: *mut Fl_Simple_Terminal, style: libc::c_int);